pyproject.toml
[tool.poetry]
name = "fertilizer-recommender"
version = "0.1.0"
description = "Clean Architecture & MLOps-ready package for Kaggle: Predicting Optimal Fertilizers (MAP@3)"
authors = ["Manda Surel <s239150.eps@gmail.com>"]
readme = "README.md"
packages = [{ include = "fertilizer_recommender", from = "src" }]
license = "MIT"
repository = "https://github.com/Manda404/fertilizer-recommender"
keywords = [
    "machine-learning",
    "mlops",
    "clean-architecture",
    "kaggle",
    "multiclass-classification",
    "map@k"
]

# -------------------------------------------------
# Runtime dependencies (INFRASTRUCTURE layer)
# -------------------------------------------------
[tool.poetry.dependencies]
python = "^3.10"

# --- Data & numerical ---
pandas = "^2.2.0"
numpy = "^1.26.0"

# --- Machine Learning ---
scikit-learn = "^1.5.0"
xgboost = "^2.0.3"
catboost = "^1.2.5"
lightgbm = "^4.6.0"

# --- Experiment tracking & reproducibility ---
mlflow = "^3.6.0"
pyyaml = "^6.0"
python-dotenv = "^1.2.1"

# --- Logging & observability ---
loguru = "^0.7.2"

# --- CLI / Presentation ---
typer = "^0.12.3"
rich = "^13.7.1"

# -------------------------------------------------
# Optional UI / API (can be removed if unused)
# -------------------------------------------------
fastapi = { version = "^0.115.0", optional = true }
uvicorn = { version = "^0.30.0", optional = true }
streamlit = { version = "^1.38.0", optional = true }

# -------------------------------------------------
# Dependency groups
# -------------------------------------------------
[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-cov = "^5.0.0"

black = "^24.0"
isort = "^5.13.0"
mypy = "^1.10.0"

pre-commit = "^3.7.0"

# --- Visualization / analysis (EDA & reports) ---
matplotlib = "^3.9.0"
seaborn = "^0.13.0"
plotly = "^6.4.0"
nbformat = "^5.10.4"

# -------------------------------------------------
# Extras (enable with: poetry install --with api,ui)
# -------------------------------------------------
[tool.poetry.extras]
api = ["fastapi", "uvicorn"]
ui = ["streamlit"]

# -------------------------------------------------
# Tooling configuration
# -------------------------------------------------
[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q --cov=src/fertilizer_recommender"

# -------------------------------------------------
# Build system
# -------------------------------------------------
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
